requirements.txt
pyrfc6266==1.0.2
python-dotenv==1.0.1
Requests==2.32.3
tqdm==4.65.0
